FEAT: realiza requisitos de 17 a 21

+++ back-end/src/database/migrations/20220921183606-sales.js
@@ -31,7 +31,7 @@ module.exports = {
       },
       total_price: {
         allowNull: false,
-        type: Sequelize.INTEGER,
+        type: Sequelize.DECIMAL(4, 2),
       },
       delivery_address: {
         allowNull: false,
